refactor(Tooltip): move to typescript
